requirements.txt
typer~=0.7.0
python-pptx
pylightxl~=1.61